+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * 【core   】     修复FileTypeUtil判断wps修改过的xlsx误判为jar的问题（pr#380@Gitee）
 * 【core   】     修复Sftp.isDir异常bug（pr#378@Gitee）
 * 【core   】     修复BeanUtil.copyProperties集合元素复制成功，读取失败的问题（issue#I41WKP@Gitee）
+* 【core   】     修复NumberChineseFormatter.chineseToNumber十位数错误（issue#1726@Gitee）
 
 -------------------------------------------------------------------------------------------------------------